pyproject.toml
[tool.poetry]
name = "django-movies"
version = "0.1.0"
description = "A django project to display movies"
authors = ["Afonso Caboz <afonso.caboz@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-split-settings = "^1.3.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
